refactor(upload): remove obsolete file validation service and streamline upload configuration
