--- frontend/src/config/mime.types
@@ -1,35 +1,0 @@
-# Format: mime/type [extensions...] # Label
-
-# Images
-image/jpeg jpg jpeg # Image JPEG
-image/png png # Image PNG
-image/gif gif # Image GIF
-image/webp webp # Image WebP
-image/bmp bmp # Image BMP
-image/svg+xml svg # Image SVG
-
-# Videos
-video/mp4 mp4 # Vidéo MP4
-video/webm webm # Vidéo WebM
-video/mpeg mpeg mpg # Vidéo MPEG
-video/quicktime mov # Vidéo QuickTime
-video/x-msvideo avi # Vidéo AVI
-
-# Documents
-application/pdf pdf # Document PDF
-text/plain txt # Fichier texte
-text/csv csv # Fichier CSV
-application/json json # Fichier JSON
-
-# MS Office
-application/msword doc # Document Word
-application/vnd.openxmlformats-officedocument.wordprocessingml.document docx # Document Word
-application/vnd.ms-excel xls # Classeur Excel
-application/vnd.openxmlformats-officedocument.spreadsheetml.sheet xlsx # Classeur Excel
-application/vnd.ms-powerpoint ppt # Présentation PowerPoint
-application/vnd.openxmlformats-officedocument.presentationml.presentation pptx # Présentation PowerPoint
-
-# Archives
-application/zip zip # Archive ZIP
-application/x-rar-compressed rar # Archive RAR
-application/x-7z-compressed 7z # Archive 7z
